+++ R/fit.R
@@ -112,20 +112,14 @@ fit.model_spec <-
     if (object$mode == "unknown") {
       rlang::abort("Please set the mode in the model specification.")
     }
-    if (!identical(class(control), class(control_parsnip()))) {
-      rlang::abort("The 'control' argument should have class 'control_parsnip'.")
-    }
+    control <- condense_control(control, control_parsnip())
     check_case_weights(case_weights, object)
 
     dots <- quos(...)
 
     if (length(possible_engines(object)) == 0) {
       prompt_missing_implementation(
-        cls = class(object)[1],
-        engine = object$engine,
-        user_specified_engine = object$user_specified_engine,
-        mode = object$mode,
-        user_specified_mode = object$user_specified_mode,
+        spec = object,
         prompt = cli::cli_abort,
         call = call2("fit")
       )
@@ -242,9 +236,8 @@ fit_xy.model_spec <-
       rlang::abort("Survival models must use the formula interface.")
     }
 
-    if (!identical(class(control), class(control_parsnip()))) {
-      rlang::abort("The 'control' argument should have class 'control_parsnip'.")
-    }
+    control <- condense_control(control, control_parsnip())
+
     if (is.null(colnames(x))) {
       rlang::abort("'x' should have column names.")
     }