+++ R/misc.R
@@ -54,46 +54,46 @@ mode_filter_condition <- function(mode, user_specified_mode) {
   rlang::quo(mode == !!mode)
 }
 
-#' @section Model Specification Checking:
-#'
-#' The helpers `spec_is_possible()`, `spec_is_loaded()`, and
-#' `prompt_missing_implementation()` provide tooling for checking
-#' model specifications. In addition to the `cls`, `engine`, and `mode`
-#' arguments, the functions take arguments `user_specified_engine` and
-#' `user_specified_mode`, denoting whether the user themselves has
-#' specified the engine or mode, respectively.
-#'
-#' `spec_is_possible()` checks against the union of
-#'
-#' * the current parsnip model environment and
-#' * the `model_info_table` of "pre-registered" model specifications
-#'
-#' to determine whether a model is well-specified. See
-#' `parsnip:::read_model_info_table()` for this table.
-#'
-#' `spec_is_loaded()` checks only against the current parsnip model environment.
-#'
-#' `spec_is_possible()` is executed automatically on `new_model_spec()`,
-#' `set_mode()`, and `set_engine()`, and `spec_is_loaded()` is executed
-#' automatically in `print.model_spec()`, among other places. `spec_is_possible()`
-#' should be used when a model specification is still "in progress" of being
-#' specified, while `spec_is_loaded` should only be called when parsnip or an
-#' extension receives some indication that the user is "done" specifying a model
-#' specification: at print, fit, addition to a workflow, or `extract_*()`, for
-#' example.
-#'
-#' When `spec_is_loaded()` is `FALSE`, the `prompt_missing_implementation()`
-#' helper will construct an informative message to prompt users to load or
-#' install needed packages. It's `prompt` argument refers to the prompting
-#' function to use, usually [cli::cli_inform] or [cli::cli_abort], and the
-#' ellipses are passed to that function.
-#'
-#' @export
-#' @keywords internal
-#' @rdname add_on_exports
-spec_is_possible <- function(cls,
-                             engine, user_specified_engine,
-                             mode, user_specified_mode) {
+# Model Specification Checking:
+#
+# The helpers `spec_is_possible()`, `spec_is_loaded()`, and
+# `prompt_missing_implementation()` provide tooling for checking
+# model specifications. In addition to the `spec`, `engine`, and `mode`
+# arguments, the functions take arguments `user_specified_engine` and
+# `user_specified_mode`, denoting whether the user themselves has
+# specified the engine or mode, respectively.
+#
+# `spec_is_possible()` checks against the union of
+#
+# * the current parsnip model environment and
+# * the `model_info_table` of "pre-registered" model specifications
+#
+# to determine whether a model is well-specified. See
+# `parsnip:::read_model_info_table()` for this table.
+#
+# `spec_is_loaded()` checks only against the current parsnip model environment.
+#
+# `spec_is_possible()` is executed automatically on `new_model_spec()`,
+# `set_mode()`, and `set_engine()`, and `spec_is_loaded()` is executed
+# automatically in `print.model_spec()`, among other places. `spec_is_possible()`
+# should be used when a model specification is still "in progress" of being
+# specified, while `spec_is_loaded` should only be called when parsnip or an
+# extension receives some indication that the user is "done" specifying a model
+# specification: at print, fit, addition to a workflow, or `extract_*()`, for
+# example.
+#
+# When `spec_is_loaded()` is `FALSE`, the `prompt_missing_implementation()`
+# helper will construct an informative message to prompt users to load or
+# install needed packages. It's `prompt` argument refers to the prompting
+# function to use, usually [cli::cli_inform] or [cli::cli_abort], and the
+# ellipses are passed to that function.
+spec_is_possible <- function(spec,
+                             engine = spec$engine,
+                             user_specified_engine = spec$user_specified_engine,
+                             mode = spec$mode,
+                             user_specified_mode = spec$user_specified_mode) {
+  cls <- class(spec)[[1]]
+
   all_model_info <-
     dplyr::full_join(
       read_model_info_table(),
@@ -115,13 +115,14 @@ spec_is_possible <- function(cls,
   return(nrow(possibilities) > 0)
 }
 
-# see ?add_on_exports for more information on usage
-#' @export
-#' @keywords internal
-#' @rdname add_on_exports
-spec_is_loaded <- function(cls,
-                           engine, user_specified_engine,
-                           mode, user_specified_mode) {
+# see notes above spec_is_possible for more information on usage
+spec_is_loaded <- function(spec,
+                           engine = spec$engine,
+                           user_specified_engine = spec$user_specified_engine,
+                           mode = spec$mode,
+                           user_specified_mode = spec$user_specified_mode) {
+  cls <- class(spec)[[1]]
+
   engine_condition <- engine_filter_condition(engine, user_specified_engine)
   mode_condition <- mode_filter_condition(mode, user_specified_mode)
 
@@ -143,9 +144,7 @@ spec_is_loaded <- function(cls,
 
 is_printable_spec <- function(x) {
   !is.null(x$method$fit$args) &&
-    spec_is_loaded(class(x)[1],
-                   x$engine, x$user_specified_engine,
-                   x$mode, x$user_specified_mode)
+    spec_is_loaded(x)
 }
 
 # construct a message informing the user that there are no
@@ -154,14 +153,15 @@ is_printable_spec <- function(x) {
 # if there's a "pre-registered" extension supporting that setup,
 # nudge the user to install/load it.
 #
-# see ?add_on_exports for more information on usage
-#' @export
-#' @keywords internal
-#' @rdname add_on_exports
-prompt_missing_implementation <- function(cls,
-                                          engine, user_specified_engine,
-                                          mode, user_specified_mode,
+# see notes above spec_is_possible for more information on usage
+prompt_missing_implementation <- function(spec,
+                                          engine = spec$engine,
+                                          user_specified_engine = spec$user_specified_engine,
+                                          mode = spec$mode,
+                                          user_specified_mode = spec$user_specified_mode,
                                           prompt, ...) {
+  cls <- class(spec)[[1]]
+
   engine_condition <- engine_filter_condition(engine, user_specified_engine)
   mode_condition <- mode_filter_condition(mode, user_specified_mode)
 
@@ -303,18 +303,17 @@ new_model_spec <- function(cls, args, eng_args, mode, user_specified_mode = TRUE
   # determine if the model specification could feasibly match any entry
   # in the union of the parsnip model environment and model_info_table.
   # if not, trigger an error based on the (possibly inferred) model spec slots.
-  if (!spec_is_possible(cls,
-                        engine, user_specified_engine,
-                        mode, user_specified_mode)) {
-    check_spec_mode_engine_val(cls, engine, mode)
-  }
-
   out <- list(
     args = args, eng_args = eng_args,
     mode = mode, user_specified_mode = user_specified_mode, method = method,
     engine = engine, user_specified_engine = user_specified_engine
   )
   class(out) <- make_classes(cls)
+
+  if (!spec_is_possible(spec = out)) {
+    check_spec_mode_engine_val(cls, engine, mode)
+  }
+
   out
 }